rename user settings js file
