+++ Resources/Private/Frontend/JavaScript/Features/UserSettings.ts
