+++ Resources/Public/JavaScript/Features.js
@@ -1,1 +1,1 @@
-(()=>{"use strict";var e={199:(e,t)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.featuresThatRequirePageReload=void 0,t.featuresThatRequirePageReload=["enba-global__enhancedIcons"]},892:function(e,t,n){var o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});const l=o(n(731)),d=o(n(905)),i=o(n(984)),c=o(n(86));"undefined"!=typeof window&&(window.addEventListener("DOMContentLoaded",(function(){window.parent!==window&&(0,i.default)()})),window.addEventListener("load",(e=>{var t;window.top===window?(0,l.default)():"typo3-contentIframe"===(null===(t=window.frameElement)||void 0===t?void 0:t.id)&&((0,d.default)(),document.querySelector(".enba-contentElementWizard__enhancedUI")&&(0,c.default)())})))},86:(e,t)=>{function n(){window.parent.document.querySelector(".t3-new-content-element-wizard-window")?window.parent.document.querySelectorAll(".t3js-media-new-content-element-wizard").forEach((function(e){var t,n;let o=null===(t=null==e?void 0:e.querySelector(".media-body"))||void 0===t?void 0:t.cloneNode(!0);if(o&&(o.removeChild(o.querySelector("br")),o.removeChild(o.querySelector("strong"))),o&&o.innerText.replace(/\n|\r|\W/g,"").length){o.classList.remove("media-body"),o.classList.add("tooltip__content");const t=document.createElement("div");t.classList.add("enba-tooltip"),t.insertAdjacentElement("afterbegin",o),t.insertAdjacentHTML("afterbegin",'<img class="tooltip__icon" src="/typo3conf/ext/enhanced_backend/Resources/Public/Icons/Info.svg" width="18" height="18" />'),null===(n=null==t?void 0:t.querySelector(".tooltip__icon"))||void 0===n||n.addEventListener("click",(function(){!function(e){e.classList.toggle("tooltip--active")}(t)})),e.insertAdjacentElement("afterbegin",t)}})):window.setTimeout(n,500)}Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(){document.querySelectorAll("typo3-backend-new-content-element-wizard-button").forEach((function(e){e.addEventListener("click",(function(){n()}))}))}},731:(e,t)=>{function n(){var e,t,d,i,c;const a=document.querySelector("#typo3-contentIframe");if(function(){var e;return"web_layout"===(null===(e=document.querySelector("#modulemenu .modulemenu-action-active"))||void 0===e?void 0:e.dataset.modulename)}()?null===(t=null===(e=document.querySelector(".t3js-scaffold-content-navigation"))||void 0===e?void 0:e.classList)||void 0===t||t.remove("content-tree--hidden"):null===(i=null===(d=document.querySelector(".t3js-scaffold-content-navigation"))||void 0===d?void 0:d.classList)||void 0===i||i.add("content-tree--hidden"),null===a)window.setTimeout(n,1e3);else{const e=a.contentDocument||(null===(c=a.contentWindow)||void 0===c?void 0:c.document),t=document.querySelector(".t3js-scaffold-content-navigation");"complete"===(null==e?void 0:e.readyState)&&t?(function(){var e,t,n;const d=document.querySelector(".t3js-scaffold-content-navigation"),i=null===(e=document.getElementById("typo3-contentIframe").contentWindow)||void 0===e?void 0:e.document,c=document.createElement("div");c.classList.add("content-tree");const a=document.createElement("div");a.classList.add("content-tree__header"),a.innerHTML='<label class="content-tree__headline">Content tree</label><i class="fa fa-solid fa-angle-down content-tree__toggle"></i>',a.onclick=l;const r=function(e){const t=e.querySelector(".typo3-docheader-pagePath + strong");if(null==t?void 0:t.innerHTML)return t.innerHTML;const n=e.querySelector(".t3js-title-inlineedit");return(null==n?void 0:n.textContent)?n.textContent:"Content Tree"}(i),u=document.createElement("div");u.classList.add("content-tree__data"),u.innerHTML='<ul><li><span class="content-tree__title">'+r+"</li></ul>",c.appendChild(a),c.appendChild(u),null==d||d.append(c);const s=null==i?void 0:i.querySelector("#PageLayoutController");if(s){const e=o(s,["t3js-page-ce-sortable"],[]);if(e)null===(t=document.querySelector(".content-tree__data .content-tree__title"))||void 0===t||t.insertAdjacentElement("afterend",e);else{const e=document.createElement("ul");e.classList.add("list-style-type--none");const t=document.createElement("li");t.innerHTML="No content on this page.",e.appendChild(t),null===(n=document.querySelector(".content-tree__data"))||void 0===n||n.insertAdjacentElement("afterend",e)}}}(),function(){var e;null===(e=document.querySelector("#typo3-contentIframe"))||void 0===e||e.addEventListener("load",(e=>{var t,o,l;if(window.top===window){const e=null===(o=null===(t=document.querySelector("#typo3-contentIframe"))||void 0===t?void 0:t.contentWindow)||void 0===o?void 0:o.document;!(null==e?void 0:e.querySelector("#EditDocumentController"))&&document.querySelector(".content-tree")&&(null===(l=document.querySelector(".content-tree"))||void 0===l||l.remove(),n())}}))}()):window.setTimeout(n,1e3)}}function o(e,t,n){var l,d,i,c,a,r,u,s,v,m,f,p,y;const h=Array.from(e.querySelectorAll("*")).filter((e=>{const n=Array.from(e.classList);return t.some((e=>null==n?void 0:n.includes(e)))}));if(0===h.length)return null;const g=document.createElement("ul");for(let e=0;e<h.length;e++){const w=h[e],S=document.createElement("li"),_=!!w.querySelector(".exampleContent > .t3-grid-container"),q=w.querySelector(".t3-page-ce-header .t3js-contextmenutrigger"),b=_?"Container":"Content element",L=(null===(l=null==w?void 0:w.querySelector('.t3-page-ce-header [data-identifier="actions-open"]'))||void 0===l?void 0:l.closest("a"))?null===(i=null===(d=null==w?void 0:w.querySelector('.t3-page-ce-header [data-identifier="actions-open"]'))||void 0===d?void 0:d.closest("a"))||void 0===i?void 0:i.cloneNode(!0):document.createElement("div");if(L&&"A"===L.tagName){for(L.innerHTML="STRONG"===(null===(a=null===(c=null==w?void 0:w.querySelector(".t3-page-ce-body .exampleContent"))||void 0===c?void 0:c.firstElementChild)||void 0===a?void 0:a.tagName)?null!==(s=null===(u=null===(r=null==w?void 0:w.querySelector(".t3-page-ce-body"))||void 0===r?void 0:r.querySelector(".exampleContent > strong"))||void 0===u?void 0:u.textContent)&&void 0!==s?s:"":"STRONG"===(null===(m=null===(v=null==w?void 0:w.querySelector(".t3-page-ce-body .exampleContent th"))||void 0===v?void 0:v.firstElementChild)||void 0===m?void 0:m.tagName)?null!==(p=null===(f=null==w?void 0:w.querySelector(".t3-page-ce-body .exampleContent th > strong"))||void 0===f?void 0:f.textContent)&&void 0!==p?p:"":b;L.classList.length>0;)L.classList.remove(L.classList.item(0)||"");L.setAttribute("target","list_frame"),L.classList.add("link-to-content-element"),q&&L.insertAdjacentElement("afterbegin",q.cloneNode(!0)),w.classList.contains("t3-page-ce-hidden")&&L.classList.add("t3-page-ce-hidden");const e=null!==(y=L.getAttribute("href"))&&void 0!==y?y:"";if(!n.includes(e)){n.push(e),S.appendChild(L);const l=o(w,t,n);l&&S.appendChild(l),g.appendChild(S)}}}return g}function l(){var e;null===(e=document.querySelector(".t3js-scaffold-content-navigation"))||void 0===e||e.classList.toggle("content-tree--collapsed")}Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(){document.querySelector(".enba-pageTree__enabledContentTree")&&n()}},984:(e,t)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(){var e,t,n,o;if("undefined"!=typeof window){const l=null===(n=null===(t=null===(e=null===window||void 0===window?void 0:window.parent)||void 0===e?void 0:e.document)||void 0===t?void 0:t.querySelector("html"))||void 0===n?void 0:n.classList,d=l?Array.from(l):[];d.filter((e=>!e.startsWith("enba"))),null===(o=document.querySelector("html"))||void 0===o||o.classList.add(...d)}}},905:(e,t,n)=>{Object.defineProperty(t,"__esModule",{value:!0});const o=n(199);function l(){const e=[];o.featuresThatRequirePageReload.forEach((t=>{var n;e.push({name:t,value:!!(null===(n=document.querySelector('[name="data['+t+']"]'))||void 0===n?void 0:n.checked)})}));const t=sessionStorage.getItem("originFeatureValues");(function(e,t){const n=e.sort(((e,t)=>e.name.localeCompare(t.name))),o=t.sort(((e,t)=>e.name.localeCompare(t.name)));for(let e=0;e<n.length;e++)if(n[e].name!==o[e].name||n[e].value!==o[e].value)return!1;return!0})(t?JSON.parse(t):[],e)||sessionStorage.setItem("reloadPage","true")}function d(){document.querySelectorAll('[name="data[enba-presets]"]').forEach((function(e){e.addEventListener("click",(function(){!function(e){switch(e){case"custom":break;case"vanilla":document.querySelectorAll('.enba-uc__feature input[type="checkbox"]').forEach((function(e){var t,n;e.checked=!!(null===(n=null===(t=e.dataset)||void 0===t?void 0:t.presets)||void 0===n?void 0:n.includes("vanilla"))}));break;case"modern":document.querySelectorAll('.enba-uc__feature input[type="checkbox"]').forEach((function(e){var t,n;e.checked=!!(null===(n=null===(t=e.dataset)||void 0===t?void 0:t.presets)||void 0===n?void 0:n.includes("modern"))}));break;case"none":document.querySelectorAll('.enba-uc__feature input[type="checkbox"]').forEach((function(e){e.checked=!1}))}}(this.getAttribute("value"))}))})),document.querySelectorAll('.enba-uc__feature input[type="checkbox"]').forEach((function(e){e.addEventListener("change",(function(){!function(e){const t=document.querySelector('[name="data[enba-presets]"][value="custom"]');t&&(t.checked=!0)}()}))}))}t.default=function(){sessionStorage.getItem("reloadPage")&&(sessionStorage.removeItem("reloadPage"),confirm("For all changes to take effect, the page must be reloaded once. Should the page be reloaded now?")&&window.parent.location.reload()),d(),document.querySelectorAll(".enba-uc-group__header").forEach((e=>{e.addEventListener("click",(function(){var t;null===(t=e.closest(".enba-uc-group"))||void 0===t||t.classList.toggle("enba-uc-group--collapsed")}))})),document.querySelectorAll('.enba-uc__feature input[type="checkbox"]').forEach((function(e){e.addEventListener("change",(function(){var t,n,l,d,i,c,a,r,u;const s=null===(t=e.getAttribute("name"))||void 0===t?void 0:t.replace(/^data\[|\]$/g,"");s&&!o.featuresThatRequirePageReload.includes(s)&&(e.checked?(null===(n=document.querySelector("html"))||void 0===n||n.classList.add(s),null===(i=null===(d=null===(l=null===window||void 0===window?void 0:window.parent)||void 0===l?void 0:l.document)||void 0===d?void 0:d.querySelector("html"))||void 0===i||i.classList.add(s)):(null===(c=document.querySelector("html"))||void 0===c||c.classList.remove(s),null===(u=null===(r=null===(a=null===window||void 0===window?void 0:window.parent)||void 0===a?void 0:a.document)||void 0===r?void 0:r.querySelector("html"))||void 0===u||u.classList.remove(s)))}))})),function(){const e=[];o.featuresThatRequirePageReload.forEach((t=>{var n;e.push({name:t,value:!!(null===(n=document.querySelector('[name="data['+t+']"]'))||void 0===n?void 0:n.checked)})})),sessionStorage.setItem("originFeatureValues",JSON.stringify(e))}(),function(){const e=document.querySelector(".btn[name='data[save]']");e&&(e.onclick=l)}()}}},t={};!function n(o){var l=t[o];if(void 0!==l)return l.exports;var d=t[o]={exports:{}};return e[o].call(d.exports,d,d.exports,n),d.exports}(892)})();
+(()=>{"use strict";var e={199:(e,t)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.featuresThatRequirePageReload=void 0,t.featuresThatRequirePageReload=["enba-global__enhancedIcons"]},892:function(e,t,n){var o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});const l=o(n(731)),d=o(n(518)),i=o(n(984)),c=o(n(86));"undefined"!=typeof window&&(window.addEventListener("DOMContentLoaded",(function(){window.parent!==window&&(0,i.default)()})),window.addEventListener("load",(e=>{var t;window.top===window?(0,l.default)():"typo3-contentIframe"===(null===(t=window.frameElement)||void 0===t?void 0:t.id)&&((0,d.default)(),document.querySelector(".enba-contentElementWizard__enhancedUI")&&(0,c.default)())})))},86:(e,t)=>{function n(){window.parent.document.querySelector(".t3-new-content-element-wizard-window")?window.parent.document.querySelectorAll(".t3js-media-new-content-element-wizard").forEach((function(e){var t,n;let o=null===(t=null==e?void 0:e.querySelector(".media-body"))||void 0===t?void 0:t.cloneNode(!0);if(o&&(o.removeChild(o.querySelector("br")),o.removeChild(o.querySelector("strong"))),o&&o.innerText.replace(/\n|\r|\W/g,"").length){o.classList.remove("media-body"),o.classList.add("tooltip__content");const t=document.createElement("div");t.classList.add("enba-tooltip"),t.insertAdjacentElement("afterbegin",o),t.insertAdjacentHTML("afterbegin",'<img class="tooltip__icon" src="/typo3conf/ext/enhanced_backend/Resources/Public/Icons/Info.svg" width="18" height="18" />'),null===(n=null==t?void 0:t.querySelector(".tooltip__icon"))||void 0===n||n.addEventListener("click",(function(){!function(e){e.classList.toggle("tooltip--active")}(t)})),e.insertAdjacentElement("afterbegin",t)}})):window.setTimeout(n,500)}Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(){document.querySelectorAll("typo3-backend-new-content-element-wizard-button").forEach((function(e){e.addEventListener("click",(function(){n()}))}))}},731:(e,t)=>{function n(){var e,t,d,i,c;const a=document.querySelector("#typo3-contentIframe");if(function(){var e;return"web_layout"===(null===(e=document.querySelector("#modulemenu .modulemenu-action-active"))||void 0===e?void 0:e.dataset.modulename)}()?null===(t=null===(e=document.querySelector(".t3js-scaffold-content-navigation"))||void 0===e?void 0:e.classList)||void 0===t||t.remove("content-tree--hidden"):null===(i=null===(d=document.querySelector(".t3js-scaffold-content-navigation"))||void 0===d?void 0:d.classList)||void 0===i||i.add("content-tree--hidden"),null===a)window.setTimeout(n,1e3);else{const e=a.contentDocument||(null===(c=a.contentWindow)||void 0===c?void 0:c.document),t=document.querySelector(".t3js-scaffold-content-navigation");"complete"===(null==e?void 0:e.readyState)&&t?(function(){var e,t,n;const d=document.querySelector(".t3js-scaffold-content-navigation"),i=null===(e=document.getElementById("typo3-contentIframe").contentWindow)||void 0===e?void 0:e.document,c=document.createElement("div");c.classList.add("content-tree");const a=document.createElement("div");a.classList.add("content-tree__header"),a.innerHTML='<label class="content-tree__headline">Content tree</label><i class="fa fa-solid fa-angle-down content-tree__toggle"></i>',a.onclick=l;const r=function(e){const t=e.querySelector(".typo3-docheader-pagePath + strong");if(null==t?void 0:t.innerHTML)return t.innerHTML;const n=e.querySelector(".t3js-title-inlineedit");return(null==n?void 0:n.textContent)?n.textContent:"Content Tree"}(i),u=document.createElement("div");u.classList.add("content-tree__data"),u.innerHTML='<ul><li><span class="content-tree__title">'+r+"</li></ul>",c.appendChild(a),c.appendChild(u),null==d||d.append(c);const s=null==i?void 0:i.querySelector("#PageLayoutController");if(s){const e=o(s,["t3js-page-ce-sortable"],[]);if(e)null===(t=document.querySelector(".content-tree__data .content-tree__title"))||void 0===t||t.insertAdjacentElement("afterend",e);else{const e=document.createElement("ul");e.classList.add("list-style-type--none");const t=document.createElement("li");t.innerHTML="No content on this page.",e.appendChild(t),null===(n=document.querySelector(".content-tree__data"))||void 0===n||n.insertAdjacentElement("afterend",e)}}}(),function(){var e;null===(e=document.querySelector("#typo3-contentIframe"))||void 0===e||e.addEventListener("load",(e=>{var t,o,l;if(window.top===window){const e=null===(o=null===(t=document.querySelector("#typo3-contentIframe"))||void 0===t?void 0:t.contentWindow)||void 0===o?void 0:o.document;!(null==e?void 0:e.querySelector("#EditDocumentController"))&&document.querySelector(".content-tree")&&(null===(l=document.querySelector(".content-tree"))||void 0===l||l.remove(),n())}}))}()):window.setTimeout(n,1e3)}}function o(e,t,n){var l,d,i,c,a,r,u,s,v,m,f,p,y;const h=Array.from(e.querySelectorAll("*")).filter((e=>{const n=Array.from(e.classList);return t.some((e=>null==n?void 0:n.includes(e)))}));if(0===h.length)return null;const g=document.createElement("ul");for(let e=0;e<h.length;e++){const w=h[e],S=document.createElement("li"),_=!!w.querySelector(".exampleContent > .t3-grid-container"),q=w.querySelector(".t3-page-ce-header .t3js-contextmenutrigger"),b=_?"Container":"Content element",L=(null===(l=null==w?void 0:w.querySelector('.t3-page-ce-header [data-identifier="actions-open"]'))||void 0===l?void 0:l.closest("a"))?null===(i=null===(d=null==w?void 0:w.querySelector('.t3-page-ce-header [data-identifier="actions-open"]'))||void 0===d?void 0:d.closest("a"))||void 0===i?void 0:i.cloneNode(!0):document.createElement("div");if(L&&"A"===L.tagName){for(L.innerHTML="STRONG"===(null===(a=null===(c=null==w?void 0:w.querySelector(".t3-page-ce-body .exampleContent"))||void 0===c?void 0:c.firstElementChild)||void 0===a?void 0:a.tagName)?null!==(s=null===(u=null===(r=null==w?void 0:w.querySelector(".t3-page-ce-body"))||void 0===r?void 0:r.querySelector(".exampleContent > strong"))||void 0===u?void 0:u.textContent)&&void 0!==s?s:"":"STRONG"===(null===(m=null===(v=null==w?void 0:w.querySelector(".t3-page-ce-body .exampleContent th"))||void 0===v?void 0:v.firstElementChild)||void 0===m?void 0:m.tagName)?null!==(p=null===(f=null==w?void 0:w.querySelector(".t3-page-ce-body .exampleContent th > strong"))||void 0===f?void 0:f.textContent)&&void 0!==p?p:"":b;L.classList.length>0;)L.classList.remove(L.classList.item(0)||"");L.setAttribute("target","list_frame"),L.classList.add("link-to-content-element"),q&&L.insertAdjacentElement("afterbegin",q.cloneNode(!0)),w.classList.contains("t3-page-ce-hidden")&&L.classList.add("t3-page-ce-hidden");const e=null!==(y=L.getAttribute("href"))&&void 0!==y?y:"";if(!n.includes(e)){n.push(e),S.appendChild(L);const l=o(w,t,n);l&&S.appendChild(l),g.appendChild(S)}}}return g}function l(){var e;null===(e=document.querySelector(".t3js-scaffold-content-navigation"))||void 0===e||e.classList.toggle("content-tree--collapsed")}Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(){document.querySelector(".enba-pageTree__enabledContentTree")&&n()}},984:(e,t)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(){var e,t,n,o;if("undefined"!=typeof window){const l=null===(n=null===(t=null===(e=null===window||void 0===window?void 0:window.parent)||void 0===e?void 0:e.document)||void 0===t?void 0:t.querySelector("html"))||void 0===n?void 0:n.classList,d=l?Array.from(l):[];d.filter((e=>!e.startsWith("enba"))),null===(o=document.querySelector("html"))||void 0===o||o.classList.add(...d)}}},518:(e,t,n)=>{Object.defineProperty(t,"__esModule",{value:!0});const o=n(199);function l(){const e=[];o.featuresThatRequirePageReload.forEach((t=>{var n;e.push({name:t,value:!!(null===(n=document.querySelector('[name="data['+t+']"]'))||void 0===n?void 0:n.checked)})}));const t=sessionStorage.getItem("originFeatureValues");(function(e,t){const n=e.sort(((e,t)=>e.name.localeCompare(t.name))),o=t.sort(((e,t)=>e.name.localeCompare(t.name)));for(let e=0;e<n.length;e++)if(n[e].name!==o[e].name||n[e].value!==o[e].value)return!1;return!0})(t?JSON.parse(t):[],e)||sessionStorage.setItem("reloadPage","true")}function d(){document.querySelectorAll('[name="data[enba-presets]"]').forEach((function(e){e.addEventListener("click",(function(){!function(e){switch(e){case"custom":break;case"vanilla":document.querySelectorAll('.enba-uc__feature input[type="checkbox"]').forEach((function(e){var t,n;e.checked=!!(null===(n=null===(t=e.dataset)||void 0===t?void 0:t.presets)||void 0===n?void 0:n.includes("vanilla"))}));break;case"modern":document.querySelectorAll('.enba-uc__feature input[type="checkbox"]').forEach((function(e){var t,n;e.checked=!!(null===(n=null===(t=e.dataset)||void 0===t?void 0:t.presets)||void 0===n?void 0:n.includes("modern"))}));break;case"none":document.querySelectorAll('.enba-uc__feature input[type="checkbox"]').forEach((function(e){e.checked=!1}))}}(this.getAttribute("value"))}))})),document.querySelectorAll('.enba-uc__feature input[type="checkbox"]').forEach((function(e){e.addEventListener("change",(function(){!function(e){const t=document.querySelector('[name="data[enba-presets]"][value="custom"]');t&&(t.checked=!0)}()}))}))}t.default=function(){sessionStorage.getItem("reloadPage")&&(sessionStorage.removeItem("reloadPage"),confirm("For all changes to take effect, the page must be reloaded once. Should the page be reloaded now?")&&window.parent.location.reload()),d(),document.querySelectorAll(".enba-uc-group__header").forEach((e=>{e.addEventListener("click",(function(){var t;null===(t=e.closest(".enba-uc-group"))||void 0===t||t.classList.toggle("enba-uc-group--collapsed")}))})),document.querySelectorAll('.enba-uc__feature input[type="checkbox"]').forEach((function(e){e.addEventListener("change",(function(){var t,n,l,d,i,c,a,r,u;const s=null===(t=e.getAttribute("name"))||void 0===t?void 0:t.replace(/^data\[|\]$/g,"");s&&!o.featuresThatRequirePageReload.includes(s)&&(e.checked?(null===(n=document.querySelector("html"))||void 0===n||n.classList.add(s),null===(i=null===(d=null===(l=null===window||void 0===window?void 0:window.parent)||void 0===l?void 0:l.document)||void 0===d?void 0:d.querySelector("html"))||void 0===i||i.classList.add(s)):(null===(c=document.querySelector("html"))||void 0===c||c.classList.remove(s),null===(u=null===(r=null===(a=null===window||void 0===window?void 0:window.parent)||void 0===a?void 0:a.document)||void 0===r?void 0:r.querySelector("html"))||void 0===u||u.classList.remove(s)))}))})),function(){const e=[];o.featuresThatRequirePageReload.forEach((t=>{var n;e.push({name:t,value:!!(null===(n=document.querySelector('[name="data['+t+']"]'))||void 0===n?void 0:n.checked)})})),sessionStorage.setItem("originFeatureValues",JSON.stringify(e))}(),function(){const e=document.querySelector(".btn[name='data[save]']");e&&(e.onclick=l)}()}}},t={};!function n(o){var l=t[o];if(void 0!==l)return l.exports;var d=t[o]={exports:{}};return e[o].call(d.exports,d,d.exports,n),d.exports}(892)})();
